stringfunc_test.go로 변경
